kernel: fair: Optimize __task_fits()
This patch is experimental.

In the previous implementation, __task_fits() always calculated
whether a task would fit in a given CPU, irrespective of whether
the CPU's capacity could accomodate it or not.So, in a case where
both the task_cpu's and the given cpu's capacities are equal, we can
assume that the task will fit.This patch therefore makes sure that if
said condition is met, __task_fits() will return true.

This change also reduced size by 60 bytes, and there was also a
noticeable decrease in hackbench scores.This change did not bring
any regressions (yet), at least on my end.

Signed-off-by: R.A.P <rhmd.19801@gmail.com>