BACKPORT, PARTIAL: random: replace custom notifier chain with standard one
commit 5acd35487dc911541672b3ffc322851769c32a56 upstream.

We previously rolled our own randomness readiness notifier, which only
has two users in the whole kernel. Replace this with a more standard
atomic notifier block that serves the same purpose with less code. Also
unexport the symbols, because no modules use it, only unconditional
builtins. The only drawback is that it's possible for a notification
handler returning the "stop" code to prevent further processing, but
given that there are only two users, and that we're unexporting this
anyway, that doesn't seem like a significant drawback for the
simplification we receive here.

Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Dominik Brodowski <linux@dominikbrodowski.net>
[Jason: for stable, also backported to crypto/drbg.c, not unexporting.]
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 705ecf13e549f924ac58d4d6b8f6358e504d8879)
[kras: only lib/vsprintf.c portion backported, due to presence of android-4.9-q
       changes introduced with commit ae89bc64b938]
Link: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/lib/vsprintf.c?h=linux-4.19.y&id=705ecf13e549f924ac58d4d6b8f6358e504d8879
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: R.A.P <rhmd.19801@gmail.com>